test: hold fp16 to a bar fp16 can actually meet
Two GPU tests failed on the first T4 run because the tolerances were guessed on CPU.

Log-probabilities were required within 0.05 of Hugging Face, but consecutive fp16 values near a GPT-2 logit are 0.06 to 0.125 apart, so the test asked for finer agreement than the format can represent; the run drifted 0.237 with 100% top-1 agreement. Raised to 0.5, with ordering still guarded by the top-1 check.

The scheduler test required batched fp16 output to track a solo run for at least five tokens, which two requests missed. Batching changes the reduction order and fp16 cannot resolve a close race, so that bar was arbitrary. It is now two tests: fp32 must match exactly, and fp16 may diverge only where the gap between the top two logits is tiny, printed either way.

Also fixes the device count printing as a control character instead of a number.

+++ src/common/device.cpp
@@ -46,7 +46,8 @@ std::string runtime_summary() {
   out << "LibTorch " << TORCH_VERSION << ", threads=" << torch::get_num_threads()
       << ", cuda=" << (torch::cuda::is_available() ? "yes" : "no");
   if (torch::cuda::is_available()) {
-    out << " (devices=" << torch::cuda::device_count() << ")";
+    // device_count() returns a narrow integer type that streams as a character, not a number.
+    out << " (devices=" << static_cast<int>(torch::cuda::device_count()) << ")";
   }
   return out.str();
 }
